Login Controller finished

+++ Marathon.API/Controllers/UsersController.cs
@@ -5,6 +5,7 @@ using System.Data.Entity;
 using System.Data.Entity.Infrastructure;
 using System.Linq;
 using System.Net;
+using System.Threading.Tasks;
 using System.Web.Http;
 using System.Web.Http.Description;
 
@@ -12,7 +13,7 @@ namespace Marathon.API.Controllers
 {
     public class UsersController : ApiController
     {
-        private MarathonDBEntities db = new MarathonDBEntities();
+        private Data.MarathonSkillsEntities db = new Data.MarathonSkillsEntities();
 
         // GET: api/Users
         public IHttpActionResult GetUsers()
@@ -28,7 +29,7 @@ namespace Marathon.API.Controllers
                     var list = new List<UserModel>();
                     foreach (var item in db.User.ToList())
                     {
-                        list.Add(new UserModel(item));
+                        list.Add(new UserModel(item,item.Role));
                     }
                     return Ok(list);
                 }
@@ -41,30 +42,38 @@ namespace Marathon.API.Controllers
         }
 
         // GET: api/Users/5
-        [ResponseType(typeof(User))]
-        public IHttpActionResult GetUser(string token, string password)
+        [ResponseType(typeof(Data.User))]
+        public async Task<IHttpActionResult> GetUser(string token, string password)
         {
             try
             {
-                var userEmail = ActiveLogins.ActiveTokens.FirstOrDefault(item => item.TokenGuid == Guid.Parse(token)).Email;
-                User user = db.User.FirstOrDefault(usert => usert.Email == userEmail);
-                if (user == null)
+                var userEmail = ActiveLogins.GetToken(Guid.Parse(token));
+                Data.User userDB = await db.User.FirstOrDefaultAsync(usert => usert.Email == userEmail.Email);
+                if (userDB == null)
                 {
                     return NotFound();
                 }
-                else if (user.Password != password)
+                else if (userDB.Password != password)
                 {
                     return Unauthorized(new System.Net.Http.Headers.AuthenticationHeaderValue("Help", "this password is incorrect"));
                 }
-                var Role = new RoleModel(user.Role);
-                var User = new UserModel(user);
-                return Ok(new { User, Role });
+                if (userEmail.AcessLevel == Enums.LevelAcess.Application)
+                {
+                    var Role = new RoleModel(userDB.Role);
+                    var User = new UserModel(userDB,userDB.Role);
+                    return Ok(User);
+                }
+                else if (userEmail.AcessLevel == Enums.LevelAcess.User)
+                {
+                    return Ok(userDB);
+                }
+                    return Unauthorized();
             }
-            catch (System.FormatException)
+            catch (FormatException)
             {
                 return BadRequest("The token format is invalid");
             }
-            catch (System.NullReferenceException)
+            catch (NullReferenceException)
             {
                 return BadRequest("This token is invalid or has already expired");
             }
@@ -73,12 +82,12 @@ namespace Marathon.API.Controllers
 
         // PUT: api/Users/5
         [ResponseType(typeof(void))]
-        public IHttpActionResult PutUser(string token, User userput, string password)
+        public async Task<IHttpActionResult> PutUser(string token, Data.User userput, string password)
         {
             try
             {
-                var userEmail = ActiveLogins.ActiveTokens.FirstOrDefault(item => item.TokenGuid == Guid.Parse(token)).Email;
-                User user = db.User.FirstOrDefault(usert => usert.Email == userEmail);
+                var userEmail = ActiveLogins.GetToken(Guid.Parse(token)).Email;
+                Data.User user = await db.User.FirstOrDefaultAsync(usert => usert.Email == userEmail);
                 if (user == null)
                 {
                     return NotFound();
@@ -98,11 +107,11 @@ namespace Marathon.API.Controllers
                 db.Entry(userput).State = EntityState.Modified;
                 try
                 {
-                    db.SaveChanges();
+                    await db.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!UserExists(userput.Email))
+                    if (!await UserExists(userput.Email))
                     {
                         return NotFound();
                     }
@@ -114,11 +123,11 @@ namespace Marathon.API.Controllers
 
                 return StatusCode(HttpStatusCode.NoContent);
             }
-            catch (System.FormatException)
+            catch (FormatException)
             {
                 return BadRequest("The token format is invalid");
             }
-            catch (System.NullReferenceException)
+            catch (NullReferenceException)
             {
                 return BadRequest("This token is invalid or has already expired");
             }
@@ -126,8 +135,8 @@ namespace Marathon.API.Controllers
         }
 
         // POST: api/Users
-        [ResponseType(typeof(User))]
-        public IHttpActionResult PostUser(User user)
+        [ResponseType(typeof(Data.User))]
+        public async Task<IHttpActionResult> PostUser(Data.User user)
         {
             if (!ModelState.IsValid)
             {
@@ -138,11 +147,11 @@ namespace Marathon.API.Controllers
 
             try
             {
-                db.SaveChanges();
+                await db.SaveChangesAsync();
             }
             catch (DbUpdateException)
             {
-                if (UserExists(user.Email))
+                if (await UserExists(user.Email))
                 {
                     return Conflict();
                 }
@@ -152,17 +161,18 @@ namespace Marathon.API.Controllers
                 }
             }
             user.Password = "this value is private";
-            return CreatedAtRoute("My", new { user.Email }, new UserModel(user));
+            return CreatedAtRoute("My", new { user.Email }, new UserModel(user,user.Role));
         }
 
         // DELETE: api/Users/5
-        [ResponseType(typeof(User))]
-        public IHttpActionResult DeleteUser(string token, string password)
+        [ResponseType(typeof(Data.User))]
+        public async Task<IHttpActionResult> DeleteUser(string token, string password)
         {
             try
             {
-                var userEmail = ActiveLogins.ActiveTokens.FirstOrDefault(item => item.TokenGuid == Guid.Parse(token)).Email;
-                User user = db.User.FirstOrDefault(usert => usert.Email == userEmail);
+                var userEmail = ActiveLogins.GetToken(Guid.Parse(token));
+
+                Data.User user = db.User.FirstOrDefault(usert => usert.Email == userEmail.Email);
                 if (user == null)
                 {
                     return NotFound();
@@ -171,16 +181,20 @@ namespace Marathon.API.Controllers
                 {
                     return Unauthorized(new System.Net.Http.Headers.AuthenticationHeaderValue("Help", "this password is incorrect"));
                 }
+                if (userEmail.AcessLevel == Enums.LevelAcess.Restricted)
+                {
+                    return Unauthorized();
+                }
 
                 db.User.Remove(user);
-                db.SaveChanges();
+                await db.SaveChangesAsync();
                 return Ok("This user has been deleted");
             }
-            catch (System.FormatException)
+            catch (FormatException)
             {
                 return BadRequest("The token format is invalid");
             }
-            catch (System.NullReferenceException)
+            catch (NullReferenceException)
             {
                 return BadRequest("This token is invalid or has already expired");
             }
@@ -196,9 +210,9 @@ namespace Marathon.API.Controllers
             base.Dispose(disposing);
         }
 
-        private bool UserExists(string id)
+        private async Task<bool> UserExists(string id)
         {
-            return db.User.Count(e => e.Email == id) > 0;
+            return await db.User.CountAsync(e => e.Email == id) > 0;
         }
     }
 }
